Refactored camera code, split scanner into additional screen
