+++ app/(tabs)/scan-receipt/_layout.tsx
@@ -1,0 +1,10 @@
+import { Stack } from "expo-router/stack";
+
+export default function ScanLayout() {
+  return (
+    <Stack>
+      <Stack.Screen name="index" options={{ headerTitle: "Camera" }} />
+      <Stack.Screen name="parser" options={{ headerTitle: "Receipt Parser" }} />
+    </Stack>
+  );
+}
